Move tests, run cargo format
