+++ src/clients/openai.rs
@@ -689,7 +689,6 @@ pub struct ChatCompletion {
     pub warnings: Vec<OrchestratorWarning>,
 }
 
-
 /// Helper to accept both string and integer for stop_reason.
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 #[serde(untagged)]
@@ -1115,5 +1114,58 @@ mod test {
         );
 
         Ok(())
+    }
+
+    /// Test deserialization of stop_reason as integer
+    #[test]
+    fn test_chat_completion_choice_stop_reason_integer() {
+        use serde_json::json;
+
+        use crate::clients::openai::{ChatCompletionChoice, StopReason};
+
+        let json_choice = json!({
+            "index": 0,
+            "message": {
+                "role": "assistant",
+                "content": "Hello!",
+                "tool_calls": [],
+                "refusal": null
+            },
+            "logprobs": null,
+            "finish_reason": "EOS_TOKEN",
+            "stop_reason": 32007
+        });
+
+        let choice: ChatCompletionChoice = serde_json::from_value(json_choice)
+            .expect("Should deserialize with integer stop_reason");
+        assert_eq!(choice.stop_reason, Some(StopReason::Integer(32007)));
+    }
+
+    /// Test deserialization of stop_reason as string
+    #[test]
+    fn test_chat_completion_choice_stop_reason_string() {
+        use serde_json::json;
+
+        use crate::clients::openai::{ChatCompletionChoice, StopReason};
+
+        let json_choice = json!({
+            "index": 0,
+            "message": {
+                "role": "assistant",
+                "content": "Hello!",
+                "tool_calls": [],
+                "refusal": null
+            },
+            "logprobs": null,
+            "finish_reason": "EOS_TOKEN",
+            "stop_reason": "32007"
+        });
+
+        let choice: ChatCompletionChoice = serde_json::from_value(json_choice)
+            .expect("Should deserialize with string stop_reason");
+        assert_eq!(
+            choice.stop_reason,
+            Some(StopReason::String("32007".to_string()))
+        );
     }
 }
