+++ tests/chat_completions_detection.rs
@@ -36,7 +36,7 @@ use fms_guardrails_orchestr8::{
         openai::{
             ChatCompletion, ChatCompletionChoice, ChatCompletionMessage, Content, ContentPart,
             ContentType, InputDetectionResult, Message, OpenAiDetections, OrchestratorWarning,
-            OutputDetectionResult, Role
+            OutputDetectionResult, Role,
         },
     },
     models::{
@@ -1486,50 +1486,4 @@ async fn orchestrator_validation_error() -> Result<(), anyhow::Error> {
     );
 
     Ok(())
-}
-
-/// Test deserialization of stop_reason as integer
-#[test]
-fn test_chat_completion_choice_stop_reason_integer() {
-    use fms_guardrails_orchestr8::clients::openai::{ChatCompletionChoice, StopReason};
-    use serde_json::json;
-
-    let json_choice = json!({
-        "index": 0,
-        "message": {
-            "role": "assistant",
-            "content": "Hello!",
-            "tool_calls": [],
-            "refusal": null
-        },
-        "logprobs": null,
-        "finish_reason": "EOS_TOKEN",
-        "stop_reason": 32007
-    });
-
-    let choice: ChatCompletionChoice = serde_json::from_value(json_choice).expect("Should deserialize with integer stop_reason");
-    assert_eq!(choice.stop_reason, Some(StopReason::Integer(32007)));
-}
-
-/// Test deserialization of stop_reason as string
-#[test]
-fn test_chat_completion_choice_stop_reason_string() {
-    use fms_guardrails_orchestr8::clients::openai::{ChatCompletionChoice, StopReason};
-    use serde_json::json;
-
-    let json_choice = json!({
-        "index": 0,
-        "message": {
-            "role": "assistant",
-            "content": "Hello!",
-            "tool_calls": [],
-            "refusal": null
-        },
-        "logprobs": null,
-        "finish_reason": "EOS_TOKEN",
-        "stop_reason": "32007"
-    });
-
-    let choice: ChatCompletionChoice = serde_json::from_value(json_choice).expect("Should deserialize with string stop_reason");
-    assert_eq!(choice.stop_reason, Some(StopReason::String("32007".to_string())));
 }
